FEATURE: switch volume to be in Long type

+++ src/main/java/com/citi/group77777/model/StockIndex.java
@@ -34,7 +34,7 @@ public class StockIndex {
     private double close;
 
     @Column(nullable = false)
-    private int volume;
+    private long volume;
 
     public StockIndex() {
         this.date = LocalDate.now();
@@ -112,7 +112,7 @@ public class StockIndex {
         this.close = close;
     }
 
-    public int getVolume() {
+    public long getVolume() {
         return volume;
     }
